Updated API for Set base classes and for its derived classes.  Resolves ATK-156.
* Base class no longer has iterators or array subscript operator (operator[]). It now has the at() function for random access.
* Derived classes have iterators and subscripting, in addition to the base class api.
* Improved types in Set base class -- SetPosition for accessing an element, vs. SetIndex for the return type.  SetSize and SetPosition should match (i.e. unsigned).
* Lots of changes throughout based on modified API.

+++ src/components/meshapi/src/Utilities.hpp
@@ -7,8 +7,8 @@
 namespace asctoolkit{
 namespace meshapi{
 
-    typedef unsigned int        MeshIndexType;
-    typedef unsigned long long  MeshSizeType;
+    typedef int        MeshIndexType;
+    typedef unsigned int        MeshSizeType;
 
     class NotImplementedException{};
